Refactoring using a "RazorPages" folder.
