+++ src/OrchardCore/Orchard.Mvc.Core/RazorPages/ModularPageActionDescriptorChangeProvider.cs
@@ -1,28 +1,27 @@
 using Microsoft.AspNetCore.Mvc.Infrastructure;
 using Microsoft.AspNetCore.Mvc.Razor.Internal;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.Options;
 using Microsoft.Extensions.Primitives;
 
-namespace Orchard.Mvc
+namespace Orchard.Mvc.RazorPages
 {
     public class ModularPageActionDescriptorChangeProvider : IActionDescriptorChangeProvider
     {
-        private const string PageSearchPattern = "/Pages/**/*.cshtml";
+        private const string PageFolder = "/Pages";
+        private const string PageSearchPattern = PageFolder + "/**/*.cshtml";
+        private const string ModularPageSearchPattern = "/**" + PageSearchPattern;
 
         private readonly IFileProvider _fileProvider;
         private readonly string _searchPattern;
 
         public ModularPageActionDescriptorChangeProvider(
-            RazorTemplateEngine templateEngine,
             IRazorViewEngineFileProviderAccessor fileProviderAccessor,
             IOptions<RazorPagesOptions> razorPagesOptions)
         {
             _fileProvider = fileProviderAccessor.FileProvider;
-            var rootDirectory = razorPagesOptions.Value.RootDirectory.TrimEnd('/');
-            _searchPattern = rootDirectory + "/**" + PageSearchPattern;
+            _searchPattern = razorPagesOptions.Value.RootDirectory.TrimEnd('/') + ModularPageSearchPattern;
         }
 
         public IChangeToken GetChangeToken()
