+++ src/OrchardCore/Orchard.Mvc.Core/RazorPages/ModularPageRouteModelConvention.cs
@@ -1,16 +1,17 @@
-using System;
 using Microsoft.AspNetCore.Mvc.ApplicationModels;
 
-namespace Orchard.Mvc
+namespace Orchard.Mvc.RazorPages
 {
     public class ModularPageRouteModelConvention : IPageRouteModelConvention
     {
+        private const string PageFolder = "/Pages";
+
         public void Apply(PageRouteModel model)
         {
             foreach (var selector in model.Selectors)
             {
-                selector.AttributeRouteModel.Template = selector.AttributeRouteModel.
-                    Template.Replace("/Pages", String.Empty);
+                selector.AttributeRouteModel.Template = selector.AttributeRouteModel
+                    .Template.Replace(PageFolder + "/", "/");
             }
         }
     }
